Move execution store to store folder
