+++ src/components/interpreter-controls.tsx
@@ -21,7 +21,7 @@ import {
   stepToPrevBlock,
   stepToView,
 } from "../engine/manual/execute";
-import { executionStore as ctx } from "../api/execution";
+import { executionStore as ctx } from "../store/execution";
 import Button from "./button";
 import Select from "./select";
